bug fix, geometry type was always the same
Hello,
I've found a bug in the .fromGeoJson() method.
Basically, the geometry type was set always from only the first element in the features array.
As a result, any geometry in the feature set was considered as polygon, if, for instance, the first feature in the GeoJSON was a polygon.
Adding the obtained shapefiles as layers on arcgis online, resulted in getting only polygons from the original GeoJSON, with polylines 'closed' to become polygons as well, and disappeared points.

Decided to go further and fix the problem myself. 
As I'm a newbie with Github and Javascript 'promise' I ask you to check the new method code for compliancy.
On my computer, the method is correctly creating shapefiles, which, uploaded on arcgis as layers, are giving the expected result.

Please consider this change.

Thanks.